+++ test/type-system.ts
@@ -1,5 +1,5 @@
-import { test } from "ava"
-import { types, getSnapshot, unprotect } from "../src"
+import { test } from 'ava'
+import { types, getSnapshot, unprotect } from '../src'
 
 const createTestFactories = () => {
     const Box = types.model({
@@ -21,20 +21,20 @@ const createTestFactories = () => {
     return { Box, Square, Cube }
 }
 
-test("it should recognize a valid snapshot", t => {
+test('it should recognize a valid snapshot', t => {
     const { Box } = createTestFactories()
 
     t.deepEqual(Box.is({ width: 1, height: 2 }), true)
     t.deepEqual(Box.is({ width: 1, height: 2, depth: 3 }), true)
 })
 
-test("it should recognize an invalid snapshot", t => {
+test('it should recognize an invalid snapshot', t => {
     const { Box } = createTestFactories()
 
-    t.deepEqual(Box.is({ width: "1", height: "2" }), false)
+    t.deepEqual(Box.is({ width: '1', height: '2' }), false)
 })
 
-test("it should check valid nodes as well", t => {
+test('it should check valid nodes as well', t => {
     const { Box } = createTestFactories()
 
     const doc = Box.create()
@@ -42,7 +42,7 @@ test("it should check valid nodes as well", t => {
     t.deepEqual(Box.is(doc), true)
 })
 
-test("it should check invalid nodes as well", t => {
+test('it should check invalid nodes as well', t => {
     const { Box } = createTestFactories()
 
     const doc = Box.create()
@@ -50,13 +50,13 @@ test("it should check invalid nodes as well", t => {
     t.deepEqual(types.model({ anotherAttr: types.number }).is(doc), false)
 })
 
-test("it should do typescript type inference correctly", t => {
+test('it should do typescript type inference correctly', t => {
     const A = types.model(
         {
             x: types.number,
             y: types.maybe(types.string),
             get z(): string {
-                return "hi"
+                return 'hi'
             },
             set z(v: string) {}
         },
@@ -70,7 +70,7 @@ test("it should do typescript type inference correctly", t => {
     )
 
     // factory is invokable
-    const a = A.create({ x: 2, y: "7" })
+    const a = A.create({ x: 2, y: '7' })
     unprotect(a)
 
     // property can be used as proper type
@@ -95,7 +95,6 @@ test("it should do typescript type inference correctly", t => {
         // MANUAL TEST not ok: z: 4
         x: 3
     })
-    unprotect(b.sub)
 
     // sub fields have proper type
     b.sub.x = 4
@@ -104,14 +103,14 @@ test("it should do typescript type inference correctly", t => {
     a.y = null
 
     const zz: string = a.z
-    a.z = "test"
+    a.z = 'test'
 
     b.sub.method()
 
     t.is(true, true) // supress no asserts warning
 })
 
-test("#66 - it should accept superfluous fields", t => {
+test('#66 - it should accept superfluous fields', t => {
     const Item = types.model({
         id: types.number,
         name: types.string
@@ -119,55 +118,55 @@ test("#66 - it should accept superfluous fields", t => {
 
     t.is(Item.is({}), false)
     t.is(Item.is({ id: 3 }), false)
-    t.is(Item.is({ id: 3, name: "" }), true)
-    t.is(Item.is({ id: 3, name: "", description: "" }), true)
+    t.is(Item.is({ id: 3, name: '' }), true)
+    t.is(Item.is({ id: 3, name: '', description: '' }), true)
 
-    const a = Item.create({ id: 3, name: "", description: "bla" } as any)
+    const a = Item.create({ id: 3, name: '', description: 'bla' } as any)
     t.is((a as any).description, undefined)
 })
 
-test("#66 - it should not require defaulted fields", t => {
+test('#66 - it should not require defaulted fields', t => {
     const Item = types.model({
         id: types.number,
-        name: types.optional(types.string, "boo")
+        name: types.optional(types.string, 'boo')
     })
 
     t.is(Item.is({}), false)
     t.is(Item.is({ id: 3 }), true)
-    t.is(Item.is({ id: 3, name: "" }), true)
-    t.is(Item.is({ id: 3, name: "", description: "" }), true)
+    t.is(Item.is({ id: 3, name: '' }), true)
+    t.is(Item.is({ id: 3, name: '', description: '' }), true)
 
-    const a = Item.create({ id: 3, description: "bla" } as any)
+    const a = Item.create({ id: 3, description: 'bla' } as any)
     t.is((a as any).description, undefined)
-    t.is(a.name, "boo")
+    t.is(a.name, 'boo')
 })
 
-test("#66 - it should be possible to omit defaulted fields", t => {
+test('#66 - it should be possible to omit defaulted fields', t => {
     const Item = types.model({
         id: types.number,
-        name: "boo"
+        name: 'boo'
     })
 
     t.is(Item.is({}), false)
     t.is(Item.is({ id: 3 }), true)
-    t.is(Item.is({ id: 3, name: "" }), true)
-    t.is(Item.is({ id: 3, name: "", description: "" }), true)
+    t.is(Item.is({ id: 3, name: '' }), true)
+    t.is(Item.is({ id: 3, name: '', description: '' }), true)
 
-    const a = Item.create({ id: 3, description: "bla" } as any)
+    const a = Item.create({ id: 3, description: 'bla' } as any)
     t.is((a as any).description, undefined)
-    t.is(a.name, "boo")
+    t.is(a.name, 'boo')
 })
 
-test("#66 - it should pick the correct type of defaulted fields", t => {
+test('#66 - it should pick the correct type of defaulted fields', t => {
     const Item = types.model({
         id: types.number,
-        name: "boo"
+        name: 'boo'
     })
 
     const a = Item.create({ id: 3 })
     unprotect(a)
 
-    t.is(a.name, "boo")
+    t.is(a.name, 'boo')
     t.throws(
         () => (a.name = 3 as any),
         `[mobx-state-tree] Error while converting \`3\` to \`string\`:
@@ -175,14 +174,14 @@ value \`3\` is not assignable to type: \`string\`.`
     )
 })
 
-test("cannot create factories with null values", t => {
+test('cannot create factories with null values', t => {
     t.throws(
         () => types.model({ x: null }),
         /The default value of an attribute cannot be null or undefined as the type cannot be inferred. Did you mean `types.maybe\(someType\)`?/
     )
 })
 
-test("can create factories with maybe primitives", t => {
+test('can create factories with maybe primitives', t => {
     const F = types.model({
         x: types.maybe(types.string)
     })
@@ -190,16 +189,16 @@ test("can create factories with maybe primitives", t => {
     t.is(F.is(undefined as any), false)
     t.is(F.is({}), true)
     t.is(F.is({ x: null }), true)
-    t.is(F.is({ x: "test" }), true)
+    t.is(F.is({ x: 'test' }), true)
     t.is(F.is({ x: 3 }), false)
 
     t.is(F.create().x, null)
     t.is(F.create({ x: undefined }).x, null)
-    t.is(F.create({ x: "" }).x, "")
-    t.is(F.create({ x: "3" }).x, "3")
+    t.is(F.create({ x: '' }).x, '')
+    t.is(F.create({ x: '3' }).x, '3')
 })
 
-test("it is possible to refer to a type", t => {
+test('it is possible to refer to a type', t => {
     const Todo = types.model(
         {
             title: types.string
@@ -210,20 +209,20 @@ test("it is possible to refer to a type", t => {
     )
 
     function x(): typeof Todo.Type {
-        return Todo.create({ title: "test" }) as any // as any to make sure the type is not inferred accidentally
+        return Todo.create({ title: 'test' }) as any // as any to make sure the type is not inferred accidentally
     }
 
     const z = x()
     unprotect(z)
 
-    z.setTitle("bla")
-    z.title = "bla"
+    z.setTitle('bla')
+    z.title = 'bla'
     // z.title = 3 // Test manual: should give compile error
 
     t.is(true, true) // supress no asserts warning
 })
 
-test(".Type should not be callable", t => {
+test('.Type should not be callable', t => {
     const Todo = types.model(
         {
             title: types.string
@@ -236,7 +235,7 @@ test(".Type should not be callable", t => {
     t.throws(() => Todo.Type)
 })
 
-test(".SnapshotType should not be callable", t => {
+test('.SnapshotType should not be callable', t => {
     const Todo = types.model(
         {
             title: types.string
@@ -249,22 +248,22 @@ test(".SnapshotType should not be callable", t => {
     t.throws(() => Todo.SnapshotType)
 })
 
-test("types instances with compatible snapshots should not be interchangeable", t => {
+test('types instances with compatible snapshots should not be interchangeable', t => {
     const A = types.model(
-        "A",
+        'A',
         {},
         {
             doA() {}
         }
     )
     const B = types.model(
-        "B",
+        'B',
         {},
         {
             doB() {}
         }
     )
-    const C = types.model("C", {
+    const C = types.model('C', {
         x: types.maybe(A)
     })
 
@@ -290,7 +289,7 @@ test("types instances with compatible snapshots should not be interchangeable", 
     }, /value of type B: <B@<root>> is not assignable to type: `A | null`/)
 })
 
-test("it handles complex types correctly", t => {
+test('it handles complex types correctly', t => {
     const Todo = types.model(
         {
             title: types.string
@@ -322,7 +321,7 @@ test("it handles complex types correctly", t => {
     t.is(true, true) // supress no asserts warning
 })
 
-test("it should provide detailed reasons why the value is not appicable", t => {
+test('it should provide detailed reasons why the value is not appicable', t => {
     const Todo = types.model(
         {
             title: types.string
@@ -355,13 +354,13 @@ test("it should provide detailed reasons why the value is not appicable", t => {
         () =>
             Store.create({
                 todos: {
-                    "1": {
+                    '1': {
                         title: true,
-                        setTitle: "hello"
+                        setTitle: 'hello'
                     }
                 },
                 amount: 1,
-                getAmount: "hello"
+                getAmount: 'hello'
             }),
         `[mobx-state-tree] Error while converting \`{"todos":{"1":{"title":true,"setTitle":"hello"}},"amount":1,"getAmount":"hello"}\` to \`AnonymousModel\`:
 at path "/todos/1/title" value \`true\` is not assignable to type: \`string\`.